move file.ts into /src/base/common
