+++ src/base/node/io.ts
@@ -3,7 +3,7 @@ import { Abortable } from 'events';
 import * as fs from 'fs';
 import * as Path from 'path';
 import { nameIncludeCheckWithRule, getFileType, pathJoin } from 'src/base/common/string';
-import { FileType } from 'src/base/node/file';
+import { FileType } from 'src/base/common/file';
 import { FileNode } from 'src/base/node/fileTree';
 
 /*******************************************************************************
